+++ CHANGELOG.md
@@ -28,6 +28,7 @@
   the next time the Debug dialog is used. Kind of a default project.
 * RR: When using the Debug dialog, check a couple standard locations for
   the RR trace directory.
+* ObjectiveC: Add support for printing ObjectiveC object via a '(objc)' pretext.
 
 ## [2.5] - 2024-12-24
 * Console now supports a subset of ANSI color codes.